ci-operator multi-stage: create service account
This addresses two problems: allowing the `secret-wrapper` to update its
Secret and allowing tests to manipulate a limited set of objects in the
test namespace.

Template tests are "allowed" to create any kind of resource because they
are evaluated by the ci-operator ServiceAccount, an admin in the test
namespace.  Multi-stage tests currently use the same account as other
tests (`default`), so their set of allowed operations is much more
limited.

This commit creates:

- a ServiceAccount specific to each test
- a Role that can
  - list/create RoleBindings
  - get/list/create the test-specific Secret
  - get imagestream/layers
- a RoleBinding that assigns the Role above and the
  `system:image-puller` ClusterRole to the test's ServiceAccount