Added a libvirt profile for remote s390x testing.
This patch was modeled after #425, so it tightly couples the architecture to the cluster profile.
Some consideration was made to leaving this profile as a generic "libvirt" profile, but
the current x86_64 tests use the AWS profile. Also, it would be useful in terms of tracking
metrics to ensure that Z/s390x can be filtered out.