Add the CVP trigger tool skeleton
CVP Trigger tool will be used by the CVP pipeline to parametrize and
trigger the verification jobs for optional operator artifacts built
internally in RH.

Right now the tool reads Prow config, finds a specified periodic job
specification, creates a Prowjob, submits it to the cluster and waits
until the job finishes.

In the future we expect this tool to perform some job parametrization.