ci-operator multi-stage: default images to `stable`
Contrary to all other areas of `ci-operator`, multi-stage tests need to
be able to use images from both the `pipeline` and `stable`
ImageStreams.

The set of images that are part of the `pipeline` IS can be identified
from the configuration, but `stable` is populated at runtime after steps
are instantiated.  Crucially, support for arbitrary external images
(i.e. outside the release configuration) is not desired, so the simple
rule of identifying if an image is declared in the configuration and
defaulting it to `stable` otherwise can be used.