rehearse: infer ci-operator config from parameters when needed
Improve pj-rehearse to try harder in cases when it needs to rehearse a
job (usually a periodic) which is not apparently tied to any specific
ci-operator config, but is configured to run `ci-operator` with explicit
`--org/--repo/--branch/--variant` options. We were already removing
these options, so we can as well save their values, and use them in case
when the normally used ones are not available (previously we would
simply fail with a "ci-operator config file --.yaml was not found"
message).

The method that removes the args from containers running `ci-operator`
was rewritten to better accomodate the need to save the values,
otherwise the changes are quite simple.