rehearse: fix excessive rehearsals for variants
The `Info.JobName()` method did not return a name including a variant
(although a variant is one of the member field of the struct, which
made rehearsal mistakenly treat more jobs as affected by a variant
ci-operator config change. By not matching job name including a variant,
it also selected all non-variant jobs as affected.

Fix `JobName()` together with Prowgen, which was actually depending on
this incorrect behavior, and appended the variant itself. Write few new
GoDocs and tests.