rehearse: set no-fail=false to stage its removal
I want to entirely remove this code (it was only useful at the beginning
when we wanted to run in a stealth mode) and to do that, I need to
remove the option from the jobs that calls `pj-rehearse`. To do that, I
need to make the false the default.