Revert "fix job name safe env var for the variant jobs"
This reverts commit 7a4efaafe3cb29db765c791965aa2834f1616573.
This reverts commit fb7981adfcac1401f3466369af1c3a80ace1d7bb.

Variants should not be in the same namespace by design (they are
different versions of the same config and we don't support config
in the same namespace). We should identify whatever the fix is
outside, because a variant using e2e-gcp should look exactly like
the core branch using e2e-gcp.