ci-operator: annotate errors at the top-most levels
In order to be able to expose a history over time of known, annotated
reasons for failure in the ci-operator process, we need a mechanism to
communicate these errors, with annotations, thoughout the codebase. In
order to allow administrators to apply default policy while also
allowing for specific error causes, a concatentating mechanism is used
in defining the error reason string. This will allow for queries for a
specific subsystem or a symptom that may occur across subsystems by
using regular expression matching on these error strings.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>