deprovision a libvirt cluster
this is for the p system currently testing the 4.3 releases. the plan is
to be able to set additional ips if we have different bastions. we could
add an outer for-loop for that in the future if needed.

the journaclctl way of checking for age/existence in time is there because
there is no way to get the age of a libvirt resource unless metadata was
added to the resource upon creation.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>