ci-operator: record org, repo branch and variant in config
We want downstream users of a CI Operator configuration that may not be
loading it from the release repository to be able to determine what org,
repo, branch and variant the configuration is for. We add these fields
as auto-generated metadata to the configuration file, and ensure that
they are set correctly when determinizing. We still want there to be
just one source of truth and we do not want users to have to set them,
or for users to be able to make a mistake setting them, so the
determinizer treats the directory tree as the source of truth.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>