ci-operator: add sane defaults to ubiquitous flags
We have a number of flags whose usage is ubiquitous and does not change
ove time. We should bake these in to the code so that we can make the
generated job configurations smaller, but continue to expose the flags
so that a quick change in configuration is all that's necessary to edit
them, not a full rebuild and redeploy.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>