prowgen: add custom secret to custom test image
In some cases, we need to use secrets during e2e tests. For example,
when we pull pre-released images from a private registry. To support
this use case, the custom test image template now accepts a
`custom_secret_name` parameter that allows users to specify a name of a
secret to mount to the custom test pod.

For users who don't specify it, a default empty secret will be used and
mounted to their test pod. This will have no effects on the tests pod.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>