ci-operator: change release step inputs and outputs
We no longer try to do an image import to read the SHA of the input
image when we have a literal image given to us by a user. This feature
relied on a bug -- creating the import in an unrelated namespace -- and
could not continue to exist. Furthermore, this feature is of limited
utility: it will only come into use when the user provides a floating
tag, not a full pull-spec, *and* two jobs with the same name and config
race to run at the same time, but resolve that tag to different SHAs.

We also no longer rely on the stable stream existing when we're running
an inport from an environment variable. It's not clear why this was the
case at all in the past, as the entire point of the input env vars was
to override the default behavior and cause an import.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>