ci-operator: don't wait for release images in templates
A template step already knows how to depend on the steps which provide
it with the paramters it needs. Therefore, today templates end up
requiring the release payload steps that create `$RELEASE_IMAGE_foo`.
These steps, in turn, require the suite of release images to be present
if necessary. There is no utlity in requiring those images, again, from
the template itself.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>