ci-operator: split the import and creation of releases
The creation of a new release payload from local image sources in an
image stream and the import of an external release are simply two
separate actions, which are executed in entirely unrelated ways. This
commit breaks the previous step that conditionally did both into two,
and chooses which to add to the graph as appropriate. This should be a
no-op for users.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>