Use methods for getting/setting the namespace on jobConfig
Before, the namespace was a public property of the jobSpec. This lead to
semi-common wrong usage where it was evaluated before it was set,
resulting in bugs.

This PR changes jobSpec to provide getters and setters instead. The
getter will log an error if the namespace was unset, as this is always a
bug. This allows us to easely identify such wrong usages without making
any tests fail due to a potentially non-fatal issue.

Also fixes a bug where the namespace property was used wrongly as
described above.