ci-operator: allow waiting before importing a release
Today, some jobs are using the following workflow:

 - build an image or set `tag_specification`
 - also provide `$RELEASE_IMAGE_LATEST`

With the previous implementation, two separate steps would create images
to fill out the stable stream and overwrite each other's work. The
effect was that the jobs could end up with images in the stable stream
that were not in the imported release from the env var, as they would be
created by the image steps and not be overwritten.

By allowing an explicit configuration for the import step, we can wait
on these other steps in the cases where we'd do in the past and continue
to provide the same user experience. This is not ideal and we need to
provide a clear way for users to simply import images they are
interested in, since the combination of `tag_specification` and a
`$RELEASE_IAMGE_LATEST` is nonsensical, as they are nominally both means
to configure which release of OpenShift is being tested.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>