config-brancher: always update all affected versions
The config-brancher tool works across a number of different versions:
 - the current version devs are working on
 - future versions we've created in advance
 - the version we are bumping the dev version to

Originally, you only had to specify a version in one of these three
fields to ensure that the config for that version was updated when the
tool was run. At some point that regressed, so for instance when you
configured a current version, future versions and a bump version, unless
you specifically listed the bump version as a future version, the bump
version would not be updated. This is problematic as when we do a bump,
the dev configuration is changed to the bump version, and the branch for
that bump version must also be changed to no longer publish images, or
we'd have two branches racing to publish to the same image tag. This
change fixes the bug and adds in a set of integration tests to verify
it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>