Registry replacer: Correctly handle From directive
We set images.*.from to point to pipelineimagestream:from. Since
the pipelineimagestream is guaranteed to be local, we do not have to add
another replacement.

In order to correctly replace remaining FROM directives in the
Dockerfile, we copy the replacing logic from openshift/builder and run
it as a pre-step before looking what additional replacements are needed.