rehearse: rework the handling of temporary CMs
In some cases rehearsals need to create temporary ConfigMaps, and their
handling was previously quite fragile. We tried to infer the ConfigMap
names from the paths and filenames but that was buggy.

The new approach is to reduce assumptions and guesswork. We only assume
the general locations of templates and cluster profiles, and use
config-updater config to determine the names of the CMs involved,
without any guessing.

Additionally, the information about CMs was used by different methods
for different purposes, which lead to some code bloat and unnecessary
complexity. This was reduced by the new `rehearse.ConfigMaps` struct
holding different facets of CM-related data, and only the necessary data
is passed to various methods. For example, the job selection methods now
only receive a set of production CM names, the job configuration methods
receive the mappings between production and rehearsal CMs, etc.

One notable change is the selection of jobs affected by template change:
previously, the code used the filename as a signifier of a job using a
template, but did not care if that filename came from an entirely
different ConfigMap, which is technically incorrect. The code was
changed to the same mechanism like cluster profile one was using:
detecting if the job uses the ConfigMap.