rehearse: unify the CM replacing for templates and cluster profiles
Now when the replacing is purely CM name-based, we can use identical
code for replacing CM references in Prowjobs. We can just search for CM
names in volume specs and replace them, and we do not need to care at
all about if the CM is a cluster profile or a template.