multi_stage: add link to registry info page for failed steps
This PR adds a link to the step registry info page for failed multistage
steps in the test logs. This makes it easier to debug failing tests.

This also makes the 404 error page for non-existent registry components
more verbose. Since it is possible for users to be running a job that
has an embedded literal step in the test config, this error page may be
hit more often so it was made a bit more verbose.