determinize-step-registry: create new tool
This PR adds a determinize-step-registry tool that performs a similar
task to the other determinize tools in this repo. It also adds a new
metadata field to the registry components that will be used by the web
UI to generate github links and get the OWNERS files for registry
components.

This PR also makes the `XSuffix` consts in the load package public and
replaces all previous hardcoded strings with those consts.