Bugzilla backporter TestRoundTrip: Deflake
When the caching transport roundtrips, it starts a go routine to use
the underlying transport and then either gets the data from the cache or
waits for the goroutine to finish. In the tests, the go routine would
sometimes be able to finish before we get to the point of checking the
cache, causing us to return data from the goroutine and failing the
test.