multi-stage: stop depending on initial release
The initial release is only created in specific scenarios. In the past,
the explicit requirement for the initial stream through an environment
link was a no-op as when no step declared the `$RELEASE_IMAGE_LATEST`
parameter, no link was added and the environment resolution did not
fail, but just set the variable to `""`. Initial releases are needed for
upgrade tests, which cannot be supported on multi-stage tests right now,
anyway, so there is no downside to removing this "dependency".

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>