ci-operator: ensure a one-to-one mapping from Dockerfile to image
We cannot allow two unique image names to be promoted after
being built from the same Dockerfile, as that would make it
impossible to take the build configuration in ocp-build-data
and determine the image build in CI that corresponds with it.
Technically, the validation we do here is more stringent --
in the most specific case it could be possible to build two
images from one Dockerfile if both were not promoted. However,
there is no real utility to that complex analysis and it's
easy and simple to enforce the stronger restriction that no
two images are built from the same Dockerfile at all.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>