ci-operator: add support for building operator index images
This PR adds support for building operator index images by adding an
`operator_index` field for images. The process is handled in a similar
manner to the bundle images, which require 2 steps, with the first step
involving building a new image to be used as a source image.

First, an image build is made that runs the `opm` command that generates
the `index.db` file and a dockerfile to build the corresponding index
image.

Second, a build is made with the previous build's image as its source,
using the dockerfile generated by `opm`.

This PR also adds validation for the `images` section of the config as
both `operator_manifests` and `operator_index` add potential for user
error.