Only enq changed imagestreamtags on imagestream update event
Currently we have two controllers that care about imagestreamtags but
watch imagestreams because imagestreamtags do not support watching. One
implication of this is that whenever any tag in a given imagestream
changes, all tags in that stream get reconciled, which is very
inefficient.

This PR adds an imagestreamtagmapper which maps imagestreams to images,
giving update events a special treatment by only enqeueing the tags that
have changed.

This mapper is then used in the promotionreconciler and the
test-images-distributor. The usage of this mapper allows the
promotionreconciler to drop request coalescing.