blocking-issue-creator: block merges to current-release too
This change makes blocking-issue-creator consistent with how other tools
(repo-brancher, mainly) treat the `--current-release` option. The
current release is considered a "future-release" and therefore blocking
issue is created for it.

Consistently, repo-brancher (the tool that performs the fast forwarding)
will fast-forward code to the release branch for the current release,
which implies the necessity to have that blocker issue.