Fix ceiling setting of ocp-build-data-enforcer
The ceiling for pr creation was used in a function that was called from
a distinct go routine that was in turn created in a loop. This means
that the order in which we process diffs is undefined, which means that
the ceiling setting doesn't do what its supposed to do, namely not
creating a PR if already $CEILING prs exist.

This PR fixes that by defering the PR creation until all processing was
done and sorting the possible PRs.