ci-operator: create Pods with resolved refs
The built-in ImageStreamTag resolution process in OpenShift is arcane
and very difficult to understand in conjuction with all of the different
levels of image imports and distribution we have. We always want to pull
from the local registry when we're creating a Pod with an ImageStreamTag
so we can just do the digest resolution ourselves and use the resolved
pull specification in the Pod's configuration. If that ends up being
pull-through with the central registry as the source, that's fine, but
opaque to the end-user ci-operator process.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>