Make sure build labels are always set correctly
Before this commit, we set the build labels like vcs ref identifier in the
projectDirectoryImageBuildStep but nowhere else. This means that all the
other images will either not have these labels or inherit them from
their base.

When such an image without correct labels gets promoted, the
promotionreconciler and possibly other automation that relies on the
correctness on them doesn't work properly anymore.