pkg/steps/lease: print resource metrics on failure
The upstream client and server errors are confusing because both
non-existent and exhausted resources result in an HTTP 404 error.

We often receive support requests because the error message (`failed to
acquire lease: resources not found`) can give the false impression that
the test infrastructure is at fault.

Sample output:

```
2020/09/04 18:08:42 Acquiring lease for "aws-quota-slice"
2020/09/04 18:08:47 warning: Failed to acquire resource, current capacity: 0 free, 1 leased
2020/09/04 18:08:47 Ran for 2m17s
error: some steps failed:
  * could not run steps: step test failed: failed to acquire lease: resources not found
```