ci-operator: validate operator.substitute.with better
`operator.substitute.with` must be resolvable to a image involved in the
ci-operation execution, otherwise we start using returning `nil`s when
computing step dependencies, leading to `nil` dereference.

Addressed in two places: on validation, and on usage (defensively, log a
warning, despite this should not happen with the validation).

Fixes: DPTP-1484