generate-registry-metadata: generate metadata individually for steps
This PR changes the way that metadata gets generated so that each step
has its own metadata file that gets stored in the same directory as the
step. This allows PRs that update metadata (like OWNERS file updates) to
be approved by the approvers of the individual steps instead of
requiring an approval by the dptp team as was the case with the
monolithic metadata file.

This also removes the separate `registry-metadata` directory and adds
the OWNERS and metadata files to the main `registry` directory. This was
done to make sure that the `configresolver` can load a registry with
metadata, which was not the case when a separate registry directory was
being used. A new make target has been added to verify that the metadata
is up-to-date and be added as a unit test to CI.