pkg/steps: Configurable activeDeadlineSeconds and terminationGracePeriodSeconds
Sometimes test steps like updates take a long time [1].  When that
happens, something wrapping the workflow may be timed out instead, and
the whole thing gets torn down without any post-test gather/teardown
steps running.  With this commit, we expose two timeout-related
properties from PodSpec [2] to step maintainers to use to control this
behavior.  Step maintainers can now set per-step timeouts that are
comfortably less than the wrapper timeout, ensuring that slow/hung
steps get killed, and the remaining steps have some time to gather
details that will explain the delay and perform other teardown.

I personally prefer matching the casing of the wrapped Kubernetes
properties, but Bruno and Petr prefer matching the snake_casing of the
existing step-config properties [3], so that's what we have in this
commit.

It seems like we'd want these to be configurable alongside all the
places where 'commands' was configurable.  But:

* I dunno what PipelineImageCacheStepConfiguration is about, so I did
  not add these properties there.
* Steve says he doesn't want them on TestStepConfiguration [4], which
  handles the 'tests' section of the configuration, including
  container, template, multi-step tests [5].

So I'm only adding them to the single-step LiteralTestStep.

[1]: https://github.com/openshift/cluster-network-operator/pull/785#issuecomment-698017421
[2]: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#podspec-v1-core
[3]: https://github.com/openshift/ci-tools/pull/1257#discussion_r494113890
[4]: https://github.com/openshift/ci-tools/pull/1257#discussion_r498932362
[5]: https://github.com/openshift/ci-tools/pull/1257#discussion_r494110156