slack-bot: handle interactions by filing issues
With this patch, the test platform Slack integration now knows how to
respond to a number of shortcut actions from users, opening dialogs that
prompt them for information in order to open issues on the DPTP Jira
board. The following types of issues are supported:
 - bugs
 - consulting requests
 - feature requests
 - incident tracking cards

Input validation is achieved with custom validation on submissiosn as
well as built-in Slack modal features; users have interactive flows to
provide them feedback on the inputs and allow them to open bugs against
this tool if we run into unexpected issues.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>