pkg/steps: Add a finalizer for pod collection
And, when we see that finalizer in the pod and are no longer
interested in watching it, remove the finalizer so the deletion can go
through.

This should avoid cases where the parallel Delete (e.g. the one in
podStep.run()) removes the pod, the change somehow slips through the
watch, and we lose track of the pod [1]:

  2020/10/06 23:04:25 error: could not wait for pod 'timeout-timeout': it is no longer present on the cluster (usually a result of a race or resource pressure. re-running the job should help)

[1]: https://github.com/openshift/ci-tools/pull/1257#issuecomment-704613892