rpm_server: Should use overwrite and not append for script creation
The use of the `>>` append redirect for the python serving script
meant that subsequent executions would not update the contents.
While this is unlikely to impact the pod, someone debugging the
script would never see their changes to the script reflected
because the new version would be appended to the python script
after a blocking function (so the updated code would never run).

Correct the typo so no other copy and paste mistakes occur.