cmd/pj-rehearse: Truncate extra rehearsals
Before this commit, pj-rehearse would just give up [1]:

  $ curl -s https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_release/12959/pull-ci-openshift-release-master-pj-rehearse/1318651994902106112/build-log.txt | tail -n2
  time="2020-10-20T20:36:51Z" level=info msg="Created a rehearsal job to be submitted" org=openshift rehearsal-job=rehearse-12959-pull-ci-openshift-kubernetes-master-e2e-cmd repo=release target-job=pull-ci-openshift-kubernetes-master-e2e-cmd target-repo=openshift/kubernetes
  time="2020-10-20T20:36:51Z" level=info msg="Would rehearse too many jobs, will not proceed" org=openshift rehearsal-jobs=68 rehearsal-threshold=45 repo=release

But skipping rehearsals for a change that touches tons of jobs, like
we have done since ef4f47ebfe (Limit rehearsals to 15 jobs,
2019-02-13, openshift/ci-operator-prowgen#78), makes it possible to
break a whole lot of things without failing a warning rehearsal.  With
this commit, we pivot from "give up and test nothing" to "test as many
of the touched jobs as we can afford".

It would be nice to intelligently truncate, e.g. if we touch a few
types of jobs, or a few different workflows.  But grouping seems
complicated, so for this commit I'm just randomly shuffling and then
dumping the tail.

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/12959/pull-ci-openshift-release-master-pj-rehearse/1318651994902106112