slack-bot: add context to job links
When a user pastes a link to a job or artifact, we often want to see the
full Spyglass link and the ci-operator configuration that gave rise to
the job in the first place. Instead of asking them for this information
we can just calculate it and post a reply.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>