WIP: cmd/ci-operator-configresolver: Expose step API
So folks can use:

  $ curl https://config.ci.openshift.org/step-registry/reference/openshift-e2e-test

and similar to get a JSON response with the rendered metadata.  The
implementation is based on a lot of copy/paste/edit out of pkg/webreg.
I'd personally rather serve this content from there, with clients
using an Accept header [1] to select between text/html and
application/json response formats:

  $ curl -H Accept:application/json https://steps.ci.openshift.org/reference/openshift-e2e-test

But Steve says that config.ci is for the API and steps.ci is just for
HTML docs and will not grow support for Accept or other response media
types.  Within that restriction, the copy/paste/edit approach allows
us to maintain as much similarity between the two path structures as
possible, so folks can easily find the API for a particular reference,
chain, workflow, etc. without having to learn a new path structure.

[1]: https://tools.ietf.org/html/rfc7231#section-5.3.2