prowgen: move code to pkg/ and simplify it
As a followup to DRYing `IsGenerated()` in
https://github.com/openshift/ci-tools/pull/1274, this change moves more
logic from `cmd/ci-operator-prowgen` to `pkg/prowgen`. Some code was
made an internal implementation detail, resulting in simpler signatures.

The labels that prowgen uses to mark (newly) generated jobs are now
internal and no longer configurable, but that was not used anyway.
`GenerateJobs` now always generates jobs with a `newly-generated` label
value, and `Prune` then switches the label to `true`.
Unfortunately many tests for `[gG]Generate*` methods passed different
labels so a lot of fixtures needed to be changed, but the actual
behavior of `ci-operator-prowgen` does not change, witnessed by
integration tests.