pkg/release: return version for official release
This PR changes the `ResolvePullSpec` function for official release to
`ResolvePullSpecAndVersion`. The `release-controller` will be importing
this package to use this function when determining releases for upgrades
and needs a tag name for the release to use as an annotation for the
upgrade job. `version` will be used as the tag name.