testgrid-config-generator: disallow ignored allow-list entries
When a job is listed as required in the `release-controller`
configuration, it will be added to a blocking dashboard and it is not
possible to overwrite that behavior with an entry in the allow-list. It
is not therefore useful to allow users to put any such jobs into the
allow-list in the first place.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>