controllers: add retries for failed images jobs (#1393)
* controllers: add retries for failed images jobs

The current image promotion reconciler ensures that we have correct
output in the form of ImageStreamTags, but misses the most common reason
that the content may be incorrect: a post-submit job building images
failed. This new controller re-triggers failed post-submit images jobs
when they are for the latest commit in the branch they're meant to
build.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>

* controller-runtime: chase upstream type changes

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>