release-job-migrator: initial commit
This PR adds a new command to ci-tools that migrates release jobs from
hand-crafted jobs to jobs generated by a ci-operator config file. There
are 2 ways it can do this conversion:
1. If a job has an `UNRESOLVED_CONFIG` set, it can identify required base images and tests from there.
2. A config file can be used to configure jobs that use the same test name.

The config file is simply a yaml map where the key is the base test
name of a job (for example, for job `release-openshift-origin-installer-e2e-gcp-upgrade-4.7`,
the base test name is `gcp-upgrade`) and the value contains 2 things:
1. The corresponding multi-stage test configuration.
2. Any extra base images that the test may require.

This is an example of the config:
```
config:
  e2e-metal-ipi:
    steps:
      cluster_profile: packet
      env:
        DEVSCRIPTS_CONFIG: |
          IP_STACK=v4
          NETWORK_TYPE=OpenShiftSDN
      workflow: baremetalds-e2e
    base_images:
      dev-scripts:
        name: test
        namespace: openshift-kni
        tag: dev-scripts
```

The tool identifies what ci-operator config the tests and images should
be added to based on the name of the periodic job. If the ci-operator
config file does not exist yet, it creates a new one with the proper release
and base image set.

Currently the tool does not support handling of jobs that upgrade
between different release streams. That will be added in a future
PR.