prowgen: provide the GCS secret to tests
We need to forward the GCS upload secret to test namespaces to allow us
to mount it into sidecar processes in test Pods. ci-operator already
knows how to do this if it's provided via flag; this change just sets
the flag and provides the data to the owning ci-operator process.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>