pkg/steps/clusterinstall/template: Get region from Boskos lease
Like openshift/release@7aa198b3c7
(ci-operator/step-registry/ipi/conf/azure: Get region from Boskos
lease, 2020-10-14, openshift/release#12584), but for AWS and GCP here
too, now that we have evidence that the approach is working.

I'm keeping a switch for AWS to give folks a pattern for selecting
zones, if AWS breaks a zone in a particular region.  We should
probably distribute that (and the shared subnets, for shared-subnet
tests?) via leases as well, but baby steps.