release-job-migrator: support upgrades
This PR adds support for migrating handcrafted upgrade release jobs
to generated upgrade release jobs. Similarly to standard release
jobs, all necessary info is inferred from the job name and all
upgrade jobs that use a standard naming scheme are supported
(there are a couple of jobs that do not and will have to be
handled manually).